unify utils
