+++ datafusion/core/src/physical_optimizer/mod.rs
@@ -26,6 +26,4 @@ pub mod enforce_distribution;
 pub mod optimizer;
 pub mod projection_pushdown;
 
-mod utils;
-
 pub use datafusion_physical_optimizer::*;
