+++ datafusion/physical-optimizer/src/lib.rs
@@ -32,5 +32,6 @@ pub mod sanity_checker;
 pub mod test_utils;
 pub mod topk_aggregation;
 pub mod update_aggr_exprs;
+pub mod utils;
 
 pub use optimizer::PhysicalOptimizerRule;
